requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.6.0
httpx==0.27.2
pydantic==2.9.2
typing-extensions==4.12.2

# OpenAI client (used for conversation loop)
openai==1.51.2

# Google Gemini SDK (optional; REST also supported)
google-generativeai==0.7.2

